add type cast

+++ dll/win32/browseui/shellfind/CFindFolder.cpp
@@ -203,8 +203,8 @@ static BOOL SearchFile(LPCWSTR lpFilePath, _SearchData *pSearchData)
         MEMORYSTATUSEX status;
         status.dwLength = sizeof(status);
         GlobalMemoryStatusEx(&status);
-        if (status.ullAvailPhys >= 2 * 1024 * 1024 * 1024) // 2GB
-            size = 2 * 1024 * 1024 * 1024 - 1;
+        if (status.ullAvailPhys >= (DWORD)(2 * 1024 * 1024 * 1024)) // 2GB
+            size = (DWORD)(2 * 1024 * 1024 * 1024 - 1);
         else
             size = (DWORD)(status.ullAvailPhys * 2 / 3); // 2/3 of physical memory
     }
